+++ storage/src/backend/rocksdb_backend.rs
@@ -2,36 +2,22 @@
 // SPDX-License-Identifier: MIT
 
 use crate::merkle_storage::{ContextValue, EntryHash};
-use crate::persistent::database::GetInMemStats;
-use crate::persistent::database::KeyValueStoreWithSchema;
-use crate::persistent::database::RocksDBStats;
-use crate::storage_backend::{StorageBackend, StorageBackendError};
+
+use crate::persistent::database::{DBError, KeyValueStoreBackend};
+use crate::storage_backend::{GarbageCollector, StorageBackendError};
 use crate::MerkleStorage;
-use rocksdb::WriteBatch;
-use rocksdb::{WriteOptions, DB};
+use rocksdb::DB;
 use serde::{Deserialize, Serialize};
 use std::ops::Deref;
 use std::sync::Arc;
 
 pub struct RocksDBBackend {
-    column_name: &'static str,
     inner: Arc<DB>,
 }
 
 impl RocksDBBackend {
-    pub fn new(db: Arc<DB>, column_name: &'static str) -> Self {
-        RocksDBBackend {
-            inner: db,
-            column_name,
-        }
-    }
-}
-
-impl RocksDBBackend {
-    fn default_write_options() -> WriteOptions {
-        let mut opts = WriteOptions::default();
-        opts.set_sync(false);
-        opts
+    pub fn new(db: Arc<DB>) -> Self {
+        RocksDBBackend { inner: db }
     }
 }
 
@@ -43,101 +29,53 @@ pub struct RocksDBBackendStats {
     cache_total: u64,
 }
 
-impl StorageBackend for RocksDBBackend {
-    fn is_persisted(&self) -> bool {
-        true
-    }
-
-    fn put(&mut self, key: &EntryHash, value: ContextValue) -> Result<bool, StorageBackendError> {
-        let cf = self.inner.cf_handle(&self.column_name).ok_or(
-            StorageBackendError::MissingColumnFamily {
-                name: &self.column_name,
-            },
-        )?;
-
-        self.inner
-            .put_cf_opt(
-                cf,
-                &key.as_ref()[..],
-                &value,
-                &Self::default_write_options(),
-            )
-            .map_err(StorageBackendError::from)
-            .map(|_| true)
-    }
-
-    fn put_batch(
-        &mut self,
-        batch: Vec<(EntryHash, ContextValue)>,
-    ) -> Result<(), StorageBackendError> {
-        let mut rocksb_batch = WriteBatch::default(); // batch containing DB key values to persist
-
-        for (k, v) in batch.iter() {
-            (self.inner.deref() as &dyn KeyValueStoreWithSchema<MerkleStorage>).put_batch(
-                &mut rocksb_batch,
-                &k,
-                &v,
-            )?;
-        }
-
-        (self.inner.deref() as &dyn KeyValueStoreWithSchema<MerkleStorage>)
-            .write_batch(rocksb_batch)?;
-
+impl GarbageCollector for RocksDBBackend {
+    fn new_cycle_started(&mut self) -> Result<(), StorageBackendError> {
         Ok(())
     }
 
-    fn merge(&mut self, key: &EntryHash, value: ContextValue) -> Result<(), StorageBackendError> {
-        let cf = self.inner.cf_handle(&self.column_name).ok_or(
-            StorageBackendError::MissingColumnFamily {
-                name: &self.column_name,
-            },
-        )?;
+    fn mark_reused(
+        &mut self,
+        _reused_keys: std::collections::HashSet<EntryHash>,
+    ) -> Result<(), StorageBackendError> {
+        Ok(())
+    }
+}
 
-        self.inner
-            .merge_cf_opt(
-                cf,
-                &key.as_ref()[..],
-                &value,
-                &Self::default_write_options(),
-            )
-            .map_err(StorageBackendError::from)
+impl KeyValueStoreBackend<MerkleStorage> for RocksDBBackend {
+    fn is_persistent(&self) -> bool {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).is_persistent()
     }
 
-    /// Warning: always returns None.
-    fn delete(&mut self, key: &EntryHash) -> Result<Option<ContextValue>, StorageBackendError> {
-        let cf = self.inner.cf_handle(&self.column_name).ok_or(
-            StorageBackendError::MissingColumnFamily {
-                name: &self.column_name,
-            },
-        )?;
-
-        self.inner
-            .delete_cf_opt(cf, &key.as_ref()[..], &Self::default_write_options())
-            .map_err(StorageBackendError::from)?;
-        Ok(None)
+    fn put(&self, key: &EntryHash, value: &ContextValue) -> Result<(), DBError> {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).put(key, value)
     }
 
-    fn get(&self, key: &EntryHash) -> Result<Option<ContextValue>, StorageBackendError> {
-        let cf = self.inner.cf_handle(&self.column_name).ok_or(
-            StorageBackendError::MissingColumnFamily {
-                name: &self.column_name,
-            },
-        )?;
-
-        let v = self
-            .inner
-            .get_cf(cf, &key.as_ref()[..])
-            .map_err(StorageBackendError::from)?;
-        Ok(v)
+    fn delete(&self, key: &EntryHash) -> Result<(), DBError> {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).delete(key)
     }
 
-    fn contains(&self, key: &EntryHash) -> Result<bool, StorageBackendError> {
-        self.get(key).map(|v| v.is_some())
+    fn merge(&self, key: &EntryHash, value: &ContextValue) -> Result<(), DBError> {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).merge(key, value)
     }
 
-    fn get_mem_use_stats(&self) -> Result<RocksDBStats, StorageBackendError> {
-        self.inner
-            .get_stats()
-            .map_err(|_| StorageBackendError::BackendError)
+    fn get(&self, key: &EntryHash) -> Result<Option<ContextValue>, DBError> {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).get(key)
+    }
+
+    fn contains(&self, key: &EntryHash) -> Result<bool, DBError> {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).contains(key)
+    }
+
+    fn retain(&self, predicate: &dyn Fn(&EntryHash) -> bool) -> Result<(), DBError> {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).retain(predicate)
+    }
+
+    fn write_batch(&self, batch: Vec<(EntryHash, ContextValue)>) -> Result<(), DBError> {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).write_batch(batch)
+    }
+
+    fn total_get_mem_usage(&self) -> Result<usize, DBError> {
+        (self.inner.deref() as &dyn KeyValueStoreBackend<MerkleStorage>).total_get_mem_usage()
     }
 }